+++ game/game_calculate.py
@@ -4,14 +4,13 @@ from game.game_cell import Cell
 
 class Calculate_word_value:
     @staticmethod
-    def calculate(cells: List[Cell]) -> int:
-        total_value = 0
-        word_multiplier = 1  
-
-        for cell in cells:
-            cell_value = cell.calculate_value()
-            if cell.multiplier_type == 'word':
-                word_multiplier *= cell.multiplier  
-            total_value += cell_value
-
-        return total_value * word_multiplier  
+    def calculate_word_value(word: list[Cell]) -> int:
+        value: int = 0
+        multiplier_word = None
+        for cell in word:
+            value = value + cell.calculate_value()
+            if cell.multiplier_type == "word" and cell.active:
+                multiplier_word = cell.multiplier
+        if multiplier_word:
+            value = value * multiplier_word
+        return value