+++ tests/test_game_calculate.py
@@ -19,7 +19,7 @@ class TestsCalculateWordValue(unittest.TestCase):
         for i, cell in enumerate(word):
             cell.add_letter(Tile(letter=letters[i], value=1))
 
-        value = Calculate_word_value.calculate(word)
+        value = Calculate_word_value.calculate_word_value(word)
         self.assertEqual(value, 4)  
 
 
@@ -34,7 +34,7 @@ class TestsCalculateWordValue(unittest.TestCase):
         for i, cell in enumerate(word):
             cell.add_letter(Tile(letter=letters[i], value=1))
 
-        value = Calculate_word_value.calculate(word)
+        value = Calculate_word_value.calculate_word_value(word)
         self.assertEqual(value, 5)
 
     def test_with_word_multiplier(self):
@@ -48,7 +48,7 @@ class TestsCalculateWordValue(unittest.TestCase):
         for i, cell in enumerate(word):
             cell.add_letter(Tile(letter=letters[i], value=1))
 
-        value = Calculate_word_value.calculate(word)
+        value = Calculate_word_value.calculate_word_value(word)
         self.assertEqual(value, 8)
 
 
@@ -59,7 +59,7 @@ class TestsCalculateWordValue(unittest.TestCase):
             Cell(letter=Tile("S", 2), multiplier=2, multiplier_type="word"),
             Cell(letter=Tile("A", 1)),
         ]
-        value = Calculate_word_value.calculate(word)
+        value = Calculate_word_value.calculate_word_value(word)
         self.assertEqual(value, 10)
 
     def test_with_letter_word_multiplier(self):
@@ -69,7 +69,7 @@ class TestsCalculateWordValue(unittest.TestCase):
             Cell(letter=Tile("S", 2), multiplier=2, multiplier_type="word"),
             Cell(letter=Tile("A", 1)),
         ]
-        value = Calculate_word_value.calculate(word)
+        value = Calculate_word_value.calculate_word_value(word)
         self.assertEqual(value, 14)
 
         
@@ -82,8 +82,8 @@ class TestsCalculateWordValue(unittest.TestCase):
         ]
         for index in range(4):
             word[index].active = False
-        value = Calculate_word_value.calculate(word)
-        self.assertEqual(value, 14)
+        value = Calculate_word_value.calculate_word_value(word)
+        self.assertEqual(value, 7)
 
     def test_with_letter_word_multiplier_no_active(self):
         word = [
@@ -92,7 +92,7 @@ class TestsCalculateWordValue(unittest.TestCase):
             Cell(letter=Tile("S", 2), multiplier=2, multiplier_type="word"),
             Cell(letter=Tile("A", 1)),
         ]
-        value = Calculate_word_value.calculate(word)
+        value = Calculate_word_value.calculate_word_value(word)
         self.assertEqual(value, 14)